requirements.txt
selenium>=4.1.3
PySimpleGUI>=4.59.0
pandas>=1.4.2
Pillow>=9.1.0
coloredlogs>=15.0.1
PyPDF2>=1.27.9
redmail>=0.4.0
XlsxWriter>=3.0.3
openpyxl>=3.0.9
python-dateutil>=2.8.2
requests>=2.27.1
openpyxl~=3.0.10
webdriver-manager~=3.8.5
bs4~=0.0.1
#pyinstaller~=5.7.0
opencv-python~=4.7.0.68
ghostscript~=0.7
Jinja2~=3.1.2
pywin32==306; sys_platform == "win32"
pywin32-ctypes~=0.2.0
selenium-stealth~=1.0.6